api/app/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Egor <wzxv_@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^21.2.0"
django = "^4.2.6"
sqlalchemy = "^2.0.22"
django-cors-headers = "^4.3.0"
djangorestframework = "^3.14.0"
pydantic = "^2.4.2"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
